Refactor code changes
