+++ mygigs_backend/backend/.env
@@ -3,8 +3,6 @@ CONSUMER_SECRET=sVo3BH4k4qI0F5BA1P85Yy5EIvAHkGj8qxIEBJuxBjox8I8Ga5wUpss4JQ7e025N
 PASSKEY=bfb279f9aa9bdbcf158e97dd71a467cd2e0c893059b10f78e6b72ada1ed2c919
 BUSINESS_SHORTCODE=174379
 MPESA_ENVIRONMENT=sandbox
-MPESA_CALLBACK_URL=https://1bc239e300a7.ngrok-free.app/core/callback/
+MPESA_CALLBACK_URL=https://01c1e2fd100b.ngrok-free.app/core/callback/
 CLERK_WEBHOOK_SECRET=whsec_MiF5K2O+8o9dkSxGT5xoG9coEHKHAQXK
 CLERK_SECRET_KEY=sk_test_0b4wLGv0hti1VSWoeLbgrywhJmLzCjAIayzuFJpvR0
-
-
