+++ mygigs_backend/backend/core/views.py
@@ -9,7 +9,7 @@ from rest_framework import parsers, serializers
 from requests.auth import HTTPBasicAuth
 from .models import MpesaTransaction, Gig, ClerkUser, Freelancer, Testimonial, Application
 from rest_framework.generics import ListAPIView, ListCreateAPIView
-from .serializers import MpesaTransactionSerializer, GigSerializer, FreelancerSerializer, TestimonialSerializer, ClerkUserSerializer, ApplicationSerializer
+from .serializers import MpesaTransactionSerializer, GigSerializer, FreelancerSerializer, TestimonialSerializer, ClerkUserSerializer
 import base64
 import json
 import os
@@ -21,9 +21,6 @@ from django.http import HttpResponse, JsonResponse
 from django.utils.decorators import method_decorator
 from django.db.models import Sum, Count
 from rest_framework.decorators import api_view, permission_classes
-from .clerck_auth import ClerkJWTAuthentication
-
-
 
 
 def get_access_token():
@@ -315,6 +312,7 @@ def clerk_webhook_handler(request):
 
     return HttpResponse(status=200)
 
+#Still does not sync user with backend
 def update_clerk_role_to_freelancer(clerk_id):
     CLERK_API_KEY = settings.CLERK_SECRET_KEY
     if not CLERK_API_KEY:
@@ -365,142 +363,6 @@ class TestimonialListCreateAPIView(generics.ListCreateAPIView):
 # class GigsListCreateAPIView(generics.ListCreateAPIView):
 #     queryset = Gig.objects.all()
 #     serializer_class = GigSerializer
-class GigCreateView(generics.CreateAPIView):
-    queryset = Gig.objects.all()
-    serializer_class = GigSerializer
-    permission_classes = [IsAuthenticated]
-    parser_classes = [parsers.MultiPartParser, parsers.FormParser]
-
-    def perform_create(self, serializer):
-        # Get the ClerkClient instance for the current authenticated user.
-        try:
-            creator_instance = ClerkUser.objects.get(clerk_id=self.request.user.clerk_id)
-        except ClerkUser.DoesNotExist:
-            raise serializers.ValidationError("Clerk user does not exist in our database.")
-            
-        # Pass the creator instance directly to the serializer's create method.
-        # This will be used by the serializer to get the email automatically.
-        serializer.save(creator=creator_instance)
-
-
-
-
-class GigListCreateAPIView(ListCreateAPIView):
-    """
-    API view to list all Gigs and create a new one.
-    """
-    queryset = Gig.objects.all()
-    serializer_class = GigSerializer
-    parser_classes = [parsers.MultiPartParser, parsers.FormParser]
-
-    permission_classes = [IsAuthenticated] 
-    def perform_create(self, serializer) -> None:
-        """
-        Saves the gig with the authenticated user as the creator.
-        """
-        try:
-            # Corrected from `ClerkClient` to `ClerkUser`
-            creator_instance = ClerkUser.objects.get(clerk_id=self.request.user.clerk_id)
-        except ClerkUser.DoesNotExist:
-            raise serializers.ValidationError("Clerk user does not exist in our database.")
-            
-        serializer.save(creator=creator_instance)
-
-
-
-# class FreelancerListCreateAPIView(ListCreateAPIView):
-#     """
-#     API view to list and create Freelancer profiles.
-#     """
-#     queryset = Freelancer.objects.all()
-#     serializer_class = FreelancerSerializer
-
-
-# class TestimonialListCreateAPIView(ListCreateAPIView):
-#     """
-#     API view to list and create Testimonials.
-#     """
-#     queryset = Testimonial.objects.all()
-#     serializer_class = TestimonialSerializer
-
-
-# @csrf_exempt
-# def clerk_webhook_handler(request) -> HttpResponse:
-#     """
-#     Handles webhooks from Clerk for user creation, update, and deletion.
-#     """
-#     print("Clerk webhook called")
-#     webhook_secret = settings.CLERK_WEBHOOK_SECRET
-#     headers = request.headers
-#     payload = request.body.decode('utf-8')
-
-#     try:
-#         wh = Webhook(webhook_secret)
-#         evt = wh.verify(payload, headers)
-#     except WebhookVerificationError:
-#         print("Webhook verification failed.")
-#         return HttpResponse(status=400)
-    
-#     event_type = evt.get("type")
-#     user_data = evt.get("data", {})
-#     clerk_id = user_data.get("id")
-
-#     if event_type in ["user.created", "user.updated"]:
-#         email = user_data.get("email_addresses", [{}])[0].get("email_address")
-#         first_name = user_data.get("first_name")
-#         last_name = user_data.get("last_name")
-
-#         try:
-#             user, created = ClerkUser.objects.get_or_create(
-#                 clerk_id=clerk_id,
-#                 defaults={
-#                     "first_name": first_name,
-#                     "last_name": last_name,
-#                     "email": email,
-#                 }
-#             )
-#             if not created:
-#                 user.first_name = first_name
-#                 user.last_name = last_name
-#                 user.email = email
-#                 user.save()
-#                 print(f"User updated: {clerk_id}")
-#             else:
-#                 print(f"User created: {clerk_id}")
-#         except Exception as e:
-#             print(f"Error in Clerk webhook user creation/update: {e}")
-
-#     elif event_type == "user.deleted":
-#         ClerkUser.objects.filter(clerk_id=clerk_id).delete()
-#         print(f"User deleted: {clerk_id}")
-
-#     return HttpResponse(status=200)
-
-def update_clerk_role_to_freelancer(clerk_id: str) -> None:
-    """
-    Calls the Clerk API to update a user's role to 'freelancer'.
-    This is an expensive, blocking call and should ideally be a background task.
-    """
-    CLERK_API_KEY = settings.CLERK_SECRET_KEY
-    if not CLERK_API_KEY:
-        print("CLERK_SECRET_KEY not set in environment.")
-        return
-
-    headers = {
-        "Authorization": f"Bearer {CLERK_API_KEY}",
-        "Content-Type": "application/json"
-    }
-    data = {
-        "public_metadata": {
-            "role": "freelancer"
-        }
-    }
-    resp = requests.patch(
-        f"https://api.clerk.com/v1/users/{clerk_id}",
-        headers=headers,
-        json=data
-    )
-    print(f"Clerk API response: {resp.status_code} {resp.text}")
 
 class DashboardDataAPIView(APIView):
     """
@@ -539,176 +401,3 @@ class DashboardDataAPIView(APIView):
                 {"error": "An internal server error occurred while fetching dashboard data."},
                 status=status.HTTP_500_INTERNAL_SERVER_ERROR
             )
-
-
-# # --- The new API View ---
-class ClerkUserListAPIView(ListAPIView):
-    """
-    API view to return a list of all ClerkUser instances.
-    """
-    # The queryset property defines the set of objects to be returned by the view.
-    # We are selecting all objects from the ClerkUser model.
-    queryset = ClerkUser.objects.all()
-    
-    serializer_class = ClerkUserSerializer
-
-
-# --- 1. Custom Permission Class ---
-
-# class IsGigCreatorOrApplicant(permissions.BasePermission):
-#     """
-#     Custom permission to only allow the Gig Creator to view/update applications
-#     or the Applicant to view their own application.
-#     """
-#     def has_object_permission(self, request, view, obj):
-#         # Read permissions are allowed to the applicant or the gig creator
-#         if request.method in permissions.SAFE_METHODS:
-#             is_applicant = obj.applicant.clerk_id == request.user.clerk_id
-#             is_gig_creator = obj.gig.creator.clerk_id == request.user.clerk_id
-#             return is_applicant or is_gig_creator
-
-#         # Write permissions (UPDATE/PATCH) are ONLY allowed to the Gig Creator
-#         if request.method in ['PUT', 'PATCH']:
-#             return obj.gig.creator.clerk_id == request.user.clerk_id
-        
-#         # Deny all other methods (like DELETE) unless explicitly defined
-#         return False
-
-
-# --- 2. Application ViewSet ---
-
-class IsApplicantOrReadOnly(permissions.BasePermission):
-    """
-    Custom permission to only allow application creators to view/edit their applications,
-    and allow gig creators/staff to view all.
-    """
-    def has_permission(self, request, view):
-        # Allow read methods (GET, HEAD, OPTIONS) for any authenticated user
-        if request.method in permissions.SAFE_METHODS and request.user.is_authenticated:
-            return True
-        # Allow POST (create application) for authenticated users
-        if view.action == 'create':
-            # This check requires request.user to be authenticated
-            return request.user.is_authenticated
-        # Fallback to general authenticated access for other methods
-        return request.user.is_authenticated
-    
-    def has_object_permission(self, request, view, obj):
-        if request.method in permissions.SAFE_METHODS:
-            return obj.applicant == request.user
-        # Allow the applicant to modify their own application (e.g., retract it)
-        return obj.applicant == request.user
-
-
-# # --- 1. Gig ViewSet (The main listing logic) ---
-class GigViewSet(viewsets.ReadOnlyModelViewSet):
-    """
-    Provides read-only access to Gig data. 
-    Handles listing gigs with filtering, searching, and price range.
-    """
-    serializer_class = GigSerializer
-    
-    # Allow anyone to view gigs, even unauthenticated users
-    permission_classes = [permissions.AllowAny] 
-    
-    def get_queryset(self):
-        # Start with all active gigs
-        queryset = Gig.objects.filter(is_active=True).order_by('-id') 
-        
-        # --- Apply Filters (Matching the frontend parameters) ---
-        search_query = self.request.query_params.get('search', None)
-        profession = self.request.query_params.get('profession', None)
-        min_price = self.request.query_params.get('min_price', None)
-        max_price = self.request.query_params.get('max_price', None)
-
-        if search_query:
-            queryset = queryset.filter(title__icontains=search_query) 
-        
-        if profession and profession != 'All':
-            queryset = queryset.filter(profession=profession)
-
-        if min_price is not None:
-            try:
-                min_price_val = float(min_price)
-                queryset = queryset.filter(price__gte=min_price_val)
-            except ValueError:
-                pass
-        
-        if max_price is not None:
-            try:
-                max_price_val = float(max_price)
-                queryset = queryset.filter(price__lte=max_price_val)
-            except ValueError:
-                pass
-
-        return queryset
-
-
-# --- 2. Application ViewSet (The application submission logic) ---
-@method_decorator(csrf_exempt, name='dispatch') 
-class ApplicationViewSet(viewsets.ModelViewSet):
-    """
-    Allows users to submit applications (POST) and view their own applications (GET).
-    """
-    serializer_class = ApplicationSerializer
-    permission_classes = [IsApplicantOrReadOnly] 
-    
-    # CRITICAL: Ensures DRF checks for a Token in the Authorization header
-    authentication_classes = [
-        authentication.TokenAuthentication,
-        authentication.SessionAuthentication,
-    ] 
-
-    def get_queryset(self):
-        # Only show applications created by the current user
-        if self.request.user.is_authenticated:
-            return Application.objects.filter(applicant=self.request.user).order_by('-id')
-        return Application.objects.none()
-
-    def perform_create(self, serializer):
-        # Automatically set the 'applicant' field to the currently logged-in user
-        serializer.save(applicant=self.request.user)
-
-
-
-class FreelancerDashboardStatsAPIView(APIView):
-    authentication_classes = [ClerkJWTAuthentication]
-    permission_classes = [IsAuthenticated]
-    def get(self, request, clerk_id):
-        freelancer = Freelancer.objects.filter(clerk_id=clerk_id).first()
-        if not freelancer:
-            return Response({"error": "Freelancer not found"}, status=404)
-
-    def get(self, request, freelancer_id):
-        total_earnings = Gig.objects.filter(freelancer_id=freelancer_id).aggregate(Sum('price'))['price__sum'] or 0
-        active_gigs = Gig.objects.filter(freelancer_id=freelancer_id, status='active').count()
-        total_clients = Application.objects.filter(gig__freelancer_id=freelancer_id).values('client_id').distinct().count()
-        completed_gigs = Gig.objects.filter(freelancer_id=freelancer_id, status='completed').count()
-        total_gigs = Gig.objects.filter(freelancer_id=freelancer_id).count()
-        completion_rate = int((completed_gigs / total_gigs) * 100) if total_gigs else 0
-
-        stats = {
-            "totalEarnings": total_earnings,
-            "activeGigs": active_gigs,
-            "totalClients": total_clients,
-            "completionRate": completion_rate,
-        }
-        return Response(stats)
-
-class FreelancerRecentApplicationsAPIView(APIView):
-    authentication_classes = [ClerkJWTAuthentication]
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, freelancer_id):
-        applications = Application.objects.filter(gig__freelancer_id=freelancer_id).order_by('-created_at')[:5]
-        serializer = ApplicationSerializer(applications, many=True)
-        return Response(serializer.data)
-
-class FreelancerGigsAPIView(APIView):
-    authentication_classes = [ClerkJWTAuthentication]
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request, freelancer_id):
-        gigs = Gig.objects.filter(freelancer_id=freelancer_id)
-        serializer = GigSerializer(gigs, many=True)
-        return Response(serializer.data)
