+++ mygigs_frontend/tsconfig.json
@@ -22,6 +22,13 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "app/payment/page.jsx", "app/admintest/page.jsx"],
+  "include": [
+    "next-env.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts",
+    "app/payment/page.jsx",
+    "app/admin-dashboard/page.jsx"
+  ],
   "exclude": ["node_modules"]
 }
